edit: deleted old files and renamed their replacements. Also changed require pathing for new file names
