+++ lib/game.rb
@@ -22,12 +22,11 @@ class Game
     @computer_submarine = Ship.new("Submarine", 2)
     @computer_board.random_placer_helper(@computer_cruiser)
     @computer_board.random_placer_helper(@computer_submarine)
-    # @computer_shot = @board.cells.keys.sample
   end
 
-  # @computer_board.place_ship.sample
-  def game_begin 
-    puts "I have laid out my ships on the grid.
+  def game_begin
+    puts 
+    "I have laid out my ships on the grid.
     You now need to lay out your two ships.
     The Cruiser is #{@cruiser.length} spaces long and the Submarine is #{@submarine.length} spaces long."
     puts @board.render(true)
@@ -49,76 +48,82 @@ class Game
 
     loop do
       puts "Now, place your submarine!"
-      # puts @board.render(true)
       user_input = gets.chomp
       formatted = user_input.upcase.delete(",").split
-        if @board.valid_placement?(@submarine, formatted) 
-          @board.place(@submarine, formatted)
-          puts "Great placement!"
-          puts @board.render(true)
-          break
-        else
-          puts "Invalid coordinates, please try again."
-          end
-          puts @board.render(true)
+      if @board.valid_placement?(@submarine, formatted) 
+        @board.place(@submarine, formatted)
+        puts "Great placement!"
+        break
+      else
+        puts "Invalid coordinates, please try again."
+      end
     end
+  end
 
+
+
+  def game_turn_start
     loop do
       puts "=============COMPUTER BOARD============="
       puts @computer_board.render(true)
       puts "==============PLAYER BOARD=============="
       puts @board.render(true)
+      if @computer_cruiser.health == 0 && @computer_submarine.health == 0
+        puts "you win!"
+        break
+      elsif @cruiser == 0 && @submarine == 0
+        puts "I win, you lose!"
+        break
+      else
+        player_turn
+        computer_turn
+      end
+    end
+  end
+  
+
+  def player_turn
+    loop do
       puts "Enter the coordinate for your shot:"
       user_input = gets.chomp
       formatted = user_input.upcase
-        if @computer_board.valid_coordinate?(formatted) && @computer_board.cells[formatted].fire_upon == true
-          if @computer_board.cells[formatted].empty? == false
-            puts "You hit my ship!"
-            puts @computer_board.render(true)
-          else
-            puts "You missed"
-            puts @computer_board.render(true)
+      if @computer_board.valid_coordinate?(formatted)   ##this is changing fired_at to true #@computer_board.cells[formatted].fired_at == true 
+        if @computer_board.cells[formatted].fired_upon? == false
+        @computer_board.cells[formatted].fire_upon
+          if @computer_board.cells[formatted].render == "M"
+            puts "Your shot on #{formatted} was a miss."
+          elsif @computer_board.cells[formatted].render == "X"
+            puts "Your shot on #{formatted} sunk my ship!"
+          elsif @computer_board.cells[formatted].render == "H"
+            puts "Your shot on #{formatted} was a hit!"
           end
-          # puts @computer_board.render(true)
-          
+          break
         else
-          puts "Please enter a valid coordinate:" 
-          break
+          puts "You hit that cell already, try again!"
         end
-      puts "I will now take my turn!"
-      @computer_shot_data = @board.cells.keys.sample
-      if @board.cells[@computer_shot_data].fired_upon? == true #loop until not fired upon
-        @computer_shot_data = @board.cells.keys.sample
-          if @board.cells[@computer_shot_data].fire_upon == true
-            @computer_shot_data = @board.cells.keys.sample
-          elsif @board.valid_coordinate(@computer_shot_data)
-            @computer_shot_data
-            puts @board.render(true)
-            computer_shot_data
-          end
-          break
-        #   @computer_shot_data = @board.cells.keys.sample
-        # if @board.valid_coordinate?(@computer_shot_data)
-        #   @computer_shot_data
-        
+      else
+        puts "Please enter a valid coordinate:" 
       end
-      
-      @computer_shot_data
-      
-      
-      if @board.cells[@computer_shot_data].render == "M"
-        puts "My shot on #{@computer_shot_data} was a miss"
-      elsif @board.cells[@computer_shot_data].render == "X"
-        puts "My shot on #{@computer_shot_data} sunk your ship!"
-      elsif @board.cells[@computer_shot_data].render == "H"
-        puts "My shot on #{@computer_shot_data} was a hit!"
-      end
-      
-      # loop do
-      # end
     end
-
-
-
   end
+
+  def computer_turn
+    puts "I will now take my turn!"
+    loop do 
+      @computer_shot_data = @board.cells.keys.sample
+      break if @board.cells[@computer_shot_data].fired_upon? == false
+    end        
+    @board.cells[@computer_shot_data].fire_upon
+    if @board.cells[@computer_shot_data].render == "M"
+      puts "My shot on #{@computer_shot_data} was a miss"
+    elsif @board.cells[@computer_shot_data].render == "X"
+      puts "My shot on #{@computer_shot_data} sunk your ship!"
+    elsif @board.cells[@computer_shot_data].render == "H"
+      puts "My shot on #{@computer_shot_data} was a hit!"
+    end
+  end
+    
+  
+      
 end
+
