--- runner_new.rb
@@ -1,11 +1,0 @@
-require './lib/ship'
-require './lib/cell'
-require './lib/board'
-require './lib/game_new.rb'
-
-@game = Game.new
-
-loop do
-  @game.main_menu
-  @game.game_turn_start
-end
